+++ common/compose/src/androidMain/kotlin/app/campfire/common/compose/layout/ResizePointerIcon.android.kt
@@ -1,0 +1,8 @@
+// Copyright 2026, Drew Heavner and the Campfire project contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+package app.campfire.common.compose.layout
+
+import androidx.compose.ui.input.pointer.PointerIcon
+
+internal actual val horizontalResizePointerIcon: PointerIcon = PointerIcon.Default